Move tomo app's page components to new routes dir
